addressed an issue with importing inside the file to be interpreted
It seems as though exec was using the namespace of enterpreter rather than that of the executed source code. As a result it couldn't see the imports. Creating a new space for it seems to have resolved it, and scoping appears to still work as expected; although I'm not entirely sure bugs related to scoping won't crop up in the future related to this.